+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 CHANGELOG
 
 ## vNext
+- Added `EnumInfoGenerator`, generating strongly typed overloads of `EnumInfo.GetName(System.Enum)`, which retrieve the name of the constant in the specified non-Flags enumeration type that has the specified value, via an allocation-free linear selection.
 
 ## v0.1.1 (2021-07-26)
 - Fixed `FriendlyNameGenerator`, now generating `Friendly.NameOf<T>()` to always return a simple name (and the containing types), rather than including namespaces when invoked with a qualified name without a using directive in scope of the call site.